+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to the "flutter-intl" extension will be documented in this f
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.3.0 - 2020-04-21
+
+- Support select messages 
+
+- Make order of supported locales consistent
+
 ## 1.2.2 - 2020-04-13
 
 - Make generated files consistent